refactor: move session module under auth umbrella
Session is an identity concept and belongs with auth::jwt, auth::oauth,
etc. No API behaviour changes — lib.rs still re-exports Session and
SessionLayer for the duration of the 0.7 branch.
