+++ src/auth/mod.rs
@@ -38,6 +38,7 @@ pub mod backup;
 pub mod jwt;
 pub mod otp;
 pub mod password;
+pub mod session;
 pub mod totp;
 
 pub mod oauth;
