+++ src/auth/session/device.rs
