+++ src/auth/session/middleware.rs
@@ -44,7 +44,7 @@ pub struct SessionLayer {
 /// # Example
 ///
 /// ```rust,no_run
-/// use modo::session::{self, SessionConfig, Store};
+/// use modo::auth::session::{self, SessionConfig, Store};
 /// use modo::cookie::{CookieConfig, key_from_config};
 /// use modo::db::Database;
 ///
