+++ tests/session_config_test.rs
@@ -1,4 +1,4 @@
-use modo::session::SessionConfig;
+use modo::auth::session::SessionConfig;
 
 #[test]
 fn test_default_values() {
