+++ tests/session_fingerprint_test.rs
@@ -1,4 +1,4 @@
-use modo::session::fingerprint::compute_fingerprint;
+use modo::auth::session::fingerprint::compute_fingerprint;
 
 #[test]
 fn fingerprint_is_64_hex() {
