+++ tests/session_meta_test.rs
@@ -1,6 +1,6 @@
 use http::HeaderMap;
+use modo::auth::session::meta::{SessionMeta, header_str};
 use modo::ip::extract_client_ip;
-use modo::session::meta::{SessionMeta, header_str};
 use std::net::IpAddr;
 
 #[test]
