+++ tests/session_store_test.rs
@@ -1,6 +1,6 @@
+use modo::auth::session::meta::SessionMeta;
+use modo::auth::session::{SessionConfig, Store};
 use modo::db::{self, ConnExt};
-use modo::session::meta::SessionMeta;
-use modo::session::{SessionConfig, Store};
 
 const CREATE_TABLE_SQL: &str = "CREATE TABLE sessions (
     id TEXT PRIMARY KEY,
